worker/requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
torch>=2.1.0
transformers==4.56.0
compressed-tensors==0.11.0
